python/requirements-kafka.txt
kafka-python==2.0.1
asyncio==3.4.3